Fixed field naming for receivedQuantityUOMCode in Line

+++ src/main/java/com/walmart/finance/ap/fds/receiving/config/DefaultValuesConfigProperties.java
@@ -44,7 +44,7 @@ public class DefaultValuesConfigProperties {
 
     private String variableWeightIndicator;
 
-    private Integer itemNumber;
+    private Long itemNumber;
 
     private Integer quantity;
